fix(server): add TLS support with self-signed certificate generation
### What

This PR wraps the server with TLS and adds functionality for generating a self-signed certificate if one doesn't exist. It checks if the existing certificate contains the local IP, generating a new one if necessary. The certificate is stored in the database, and an endpoint `/admin-api/certificate` is created to retrieve and download the certificate.

### Why

Adding TLS support ensures secure communication and enables users to access the server from external location on the same network. Generating self-signed certificates automatically simplifies the setup process. Storing the certificate in the database ensures it persists between server restarts, and the new endpoint allows easy retrieval.

### How

- Wrapped the server with TLS.
- Added functionality to generate a self-signed certificate if none exists.
- Implemented logic to check the existing certificate for the local IP and regenerate it if necessary.
- Stored the certificate in the database under ssl_certs:server key.
- Created the endpoint `/admin-api/certificate` to fetch and download the certificate.

### Related Issues

Closes https://github.com/calimero-network/admin-dashboard/issues/26